Moving service worker job related code to its own file. (#628)
b/286597506

(cherry picked from commit 6e280c8f020bb5cccdbbf070f60805ccdbfea86c)
