+++ cobalt/worker/service_worker_persistent_settings.h
@@ -48,16 +48,16 @@ class ServiceWorkerPersistentSettings {
     Options(web::WebSettings* web_settings,
             network::NetworkModule* network_module,
             web::UserAgentPlatformInfo* platform_info,
-            ServiceWorkerJobs* service_worker_jobs, const GURL& url)
+            ServiceWorkerContext* service_worker_context, const GURL& url)
         : web_settings(web_settings),
           network_module(network_module),
           platform_info(platform_info),
-          service_worker_jobs(service_worker_jobs),
+          service_worker_context(service_worker_context),
           url(url) {}
     web::WebSettings* web_settings;
     network::NetworkModule* network_module;
     web::UserAgentPlatformInfo* platform_info;
-    ServiceWorkerJobs* service_worker_jobs;
+    ServiceWorkerContext* service_worker_context;
     const GURL& url;
   };
