+++ cobalt/worker/service_worker_registration_map.cc
@@ -29,7 +29,7 @@
 #include "cobalt/script/exception_message.h"
 #include "cobalt/script/promise.h"
 #include "cobalt/script/script_value.h"
-#include "cobalt/worker/service_worker_jobs.h"
+#include "cobalt/worker/service_worker_context.h"
 #include "cobalt/worker/service_worker_registration_object.h"
 #include "cobalt/worker/service_worker_update_via_cache.h"
 #include "url/gurl.h"
@@ -252,7 +252,7 @@ bool ServiceWorkerRegistrationMap::IsUnregistered(
 }
 
 void ServiceWorkerRegistrationMap::HandleUserAgentShutdown(
-    ServiceWorkerJobs* jobs) {
+    ServiceWorkerContext* context) {
   TRACE_EVENT0("cobalt::worker",
                "ServiceWorkerRegistrationMap::HandleUserAgentShutdown()");
   DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
@@ -270,7 +270,7 @@ void ServiceWorkerRegistrationMap::HandleUserAgentShutdown(
       // active worker is null, invoke Clear Registration with registration and
       // continue to the next iteration of the loop.
       if (!registration->waiting_worker() && !registration->active_worker()) {
-        jobs->ClearRegistration(registration);
+        context->ClearRegistration(registration);
         continue;
       } else {
         // 1.1.2. Else, set installingWorker to null.
@@ -283,7 +283,7 @@ void ServiceWorkerRegistrationMap::HandleUserAgentShutdown(
       // substep in parallel:
 
       // 1.2.1. Invoke Activate with registration.
-      jobs->Activate(registration);
+      context->Activate(registration);
     }
   }
 }
